Renamed data structures directory and removed pub(crate) calls
