--- src-tauri/src/analytics/aux_data_structures/mod.rs
@@ -1,6 +1,0 @@
-pub(crate) mod binary_heap;
-pub(crate) mod cut;
-pub(crate) mod neighbor_info;
-pub(crate) mod stoer_wagner_ds;
-pub(crate) mod timeline;
-pub(crate) mod union_find;
