+++ src-tauri/src/analytics/aux_functions/mod.rs
@@ -1,4 +1,4 @@
-pub(crate) mod convert_to_graph_from_string;
-pub(crate) mod edge_factory;
-pub(crate) mod get_most_similar;
-pub(crate) mod take_snapshot;
+pub mod convert_to_graph_from_string;
+pub mod edge_factory;
+pub mod get_most_similar;
+pub mod take_snapshot;
