+++ src-tauri/src/analytics/data_structures/cut.rs
