+++ src-tauri/src/analytics/data_structures/neighbor_info.rs
