+++ src-tauri/src/analytics/data_structures/timeline.rs
