+++ src-tauri/src/constructors/init/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod init_edge_map;
-pub(crate) mod init_graph;
-pub(crate) mod init_node_map;
+pub mod init_edge_map;
+pub mod init_graph;
+pub mod init_node_map;
