+++ src-tauri/src/constructors/mock/mod.rs
@@ -1,1 +1,1 @@
-pub(crate) mod mocks;
+pub mod mocks;
