+++ src-tauri/src/data_conversion/mod.rs
@@ -1,2 +1,2 @@
-pub(crate) mod distance_constants;
-pub(crate) mod distance_conversion;
+pub mod distance_constants;
+pub mod distance_conversion;
