+++ src-tauri/src/graph/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod edge;
-pub(crate) mod graph_ds;
-pub(crate) mod node;
+pub mod edge;
+pub mod graph_ds;
+pub mod node;
